Change Socrata module to OpenData
Use find_in_batches to avoid server memory problems
