+++ README.md
@@ -60,7 +60,7 @@ SODA_DATASET_IDENTIFIER: "dataset_identifier"
 
 There are two rake tasks available to interact with this data:
 
-- `rake open_data:export:participatory_processess` to **export** the data to a file
-- `rake open_data:publish_to_socrata:participatory_processess` to **update** the remote Socrata dataset
+- `rake open_data:participatory_processes:export` to **export** the data to a CSV file
+- `rake open_data:participatory_processes:publish_to_socrata` to **update** the remote Socrata dataset
 
-The logic of these rake tasks has been extracted to a `Module` in `lib/socrata.rb`, and `ParticipatoryProcessSerializer`, located at `decidim-process-extended/app/serializers/decidim/participatory_processes`, is the `Class` responsible for transforming the objects into data for the Socrata dataset.
+The logic of these rake tasks has been extracted to a `Module` in `lib/open_data.rb`, and `ParticipatoryProcessSerializer`, located at `decidim-process-extended/app/serializers/decidim/participatory_processes`, is the `Class` responsible for transforming the objects into data for the Socrata dataset.
