+++ config/schedule.rb
@@ -15,5 +15,5 @@ every 1.day, at: '2:00 am' do
 end
 
 every :sunday, at: '11:59 pm' do
-  rake 'open_data:publish_to_socrata:participatory_processess'
+  rake 'rake open_data:participatory_processes:publish_to_socrata'
 end
