+++ decidim-process-extended/app/views/decidim/participatory_processes/admin/participatory_processes/_form.html.erb
@@ -164,11 +164,9 @@
     <div class="row column">
       <%= form.check_box :private_space, label: t(".private_space") %>
     </div>
-
     <div class="row column">
       <%= form.check_box :promoted %>
     </div>
-
     <div class="row column">
       <%= form.check_box :show_home, label: t(".appear_home") %>
     </div>
